refactor(#3038): isolate tmux watcher child tests
